Add global option `python_kit_profile`
The default behavior predicts using the release
name as a python-kit profile. We want to use a single profile
for all our branches used in MARK.
This commit introduces a new global option for `ego.conf` that
permits to override of the default logic and specifying the
profile to use for python-kit. The option is `python_kit_profile`.

```
$> cat /etc/ego.conf
[global]
release = mark-testing
python_kit_profile = mark
sync_base_url = https://github.com/macaroni-os/{repo}
```